fix: remove a long-standing bug from QuickselectH
I modeled QuickselectH too closely on QuickselectL, but the two
partitions have different properties. In particular, the Lomuto method
is single pass, but the Hoare may require multiple passes.

As a result, if I use Hoare’s partition, I need to wrap the switch
statement in a `for { .... }` loop.

TODO: clean up the tests throughout this project.